% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_eval.R
\name{poly_eval}
\alias{poly_eval}
\title{Evaluates a polynomial.}
\usage{
poly_eval(order, a, x)
}
\arguments{
\item{order}{the order of the polynomial.}

\item{a}{a vector of coefficients.}

\item{x}{the values of x.}
}
\value{
The evaluation of the polynomial.
}
\description{
Evaluates a polynomial.
}
\examples{
poly_eval(5, c(1, 2, 3, 4, 5), 5)
}
\author{
Goutham Swaminathan \email{goutham1220@gmail.com}
}
